Fix int cast was used in "getCapacityAsLong()" in "WorkerMachineBlockEntity" and "WorkerFluidMachineBlockEntity" base classes
(cherry picked from commit c3beca47a7ac32ba9b1a5b8887215c5ed59c9757)

+++ src/main/java/me/jddev0/ep/block/entity/base/WorkerFluidMachineBlockEntity.java
@@ -49,7 +49,7 @@ public abstract class WorkerFluidMachineBlockEntity<W>
         return new EnergizedPowerEnergyStorage(baseEnergyCapacity) {
             @Override
             public long getCapacityAsLong() {
-                return Math.max(1, (int)Math.ceil(capacity * upgradeModuleInventory.getModifierEffectProduct(
+                return Math.max(1, (long)Math.ceil(capacity * upgradeModuleInventory.getModifierEffectProduct(
                         UpgradeModuleModifier.ENERGY_CAPACITY)));
             }
